requirements.txt
fastapi
uvicorn
pillow
reportlab
jinja2
python-multipart